+++ deployment/deploy.sh
@@ -6,9 +6,10 @@
 npm run build
 tar -czf dist.tar.gz dist
 scp dist.tar.gz alakajam@alakajam.com:/var/www/alakajam/dist.tar.gz
-echo "Archive created."
-echo "If package.json has changed, don't forget to do a 'npm install'."
-echo "If static files have changed, don't forget to do a 'git pull'."
 
-# Manual actions on the server, as alakajam
-# > git pull && tar -xzf dist.tar.gz && rm dist.tar.gz && pm2 restart alakajam
+echo "Archive created."
+echo ""
+echo "Manual actions on the server, as the alakajam user:"
+echo "> git pull && tar -xzf dist.tar.gz && rm dist.tar.gz && pm2 restart alakajam"
+echo ""
+echo "If package.json has changed, don't forget to do a 'npm install'."